backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
openai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv